requirements.txt
pytest
requests
jsonschema
python-dotenv
allure-pytest
Faker
